+++ Source/Cypher/app/src/main/AndroidManifest.xml
@@ -65,7 +65,7 @@
         </activity>
         <activity android:name="org.artoolkit.ar.base.camera.CameraPreferencesActivity" />
         <service
-            android:name=".NetworkingService"
+            android:name=".networking.NetworkingService"
             android:exported="false"/>
     </application>
