+++ Source/Cypher/app/src/main/java/com/semaphore_soft/apps/cypher/ConnectionLobbyActivity.java
@@ -18,6 +18,7 @@ import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.semaphore_soft.apps.cypher.networking.NetworkingService;
 import com.semaphore_soft.apps.cypher.ui.PlayerID;
 import com.semaphore_soft.apps.cypher.ui.PlayerIDAdapter;
